EvenFactory is moved to Events package
